Renamed: component to SelectAttributes.vue
