+++ apps/frontend/src/components/SelectAttributes.vue
